Rework RailsParallel bootup and schema load.
We now unconditionally load the main schema.  This prevents situations
where the app loads with one database and then operates on another,
which confuses ActiveRecord column discovery.

To keep things streamlined, schema generation has been moved to the
Rake module's "launch" method.  The RP runner is launched, and as soon
as Rake gets the signal that the runner is starting, it generates and
sends the schema.  With luck, the schema is ready and sent by the time
the runner needs it.